streamlit_app/requirements.txt
streamlit==1.29.0
streamlit-ace==0.1.1
pexpect==4.8.0
python-dotenv==1.0.0
google-generativeai==0.3.1
watchdog==3.0.0
streamlit-tree-select==0.0.5
zipfile36==0.1.3
pyyaml==6.0.1
# Removed firebase-admin and authentication dependencies

